+++ src/zscanner/scanner_functions.c
@@ -14,6 +14,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <config.h>
 #include "zscanner/scanner_functions.h"
 
 #include <stdint.h>
@@ -752,7 +753,8 @@ int date_to_timestamp(uint8_t *buff, uint32_t *timestamp)
 	if (is_leap_year[year]) {
 		if (month > 2) {
 			leap_day = 1; // Add one day in case of leap year.
-		} else if (month == 2 && day > (days_in_months[month] + 1)) {
+		} else if (month == 2 &&
+		           day > (uint32_t)(days_in_months[month] + 1)) {
 			return ZSCANNER_EBAD_DATE;
 		}
 	} else if (day > days_in_months[month]){
@@ -981,4 +983,3 @@ int find_rdata_blocks(scanner_t *s)
 		return KNOT_EOK;
 	}
 }
-
